requirements.txt
django==1.4
PIL
django-debug-toolbar
django-extensions
django-haystack
django-social-auth
django-storages
django_compressor
gunicorn
markdown
psycopg2
python-dateutil==1.5
python-linkedin
requests
slimit
south
whoosh
yql
-e git+https://github.com/sunlightlabs/python-transparencydata.git#egg=python_transparencydata-dev
-e git+https://github.com/drinks/python-inboxinfluence.git#egg=python_inboxinfluence-dev
-e git+https://github.com/drinks/python-crunchbase.git#egg=python_crunchbase-dev
-e git+https://github.com/liberation/django-jsonfield.git#egg=django_jsonfield-dev
-e git+https://github.com/jschrewe/django-genericadmin#egg=genericadmin
# important to stick with this fork; relies on a new model field
-e git+https://github.com/drinks/django-postmark.git#egg=django_postmark-dev